requirements.txt
python-telegram-bot[job-queue]==21.*
python-dotenv
flask
requests
beautifulsoup4
